Moved question generator and pieces to questions folder
