+++ code_tests/low_cost_or_live_api_tests/test_question_generator.py
@@ -3,9 +3,13 @@ from datetime import datetime, timedelta
 
 import pytest
 
-from forecasting_tools.agents_and_tools.question_generator import (
+from forecasting_tools.agents_and_tools.question_generators.generated_question import (
     GeneratedQuestion,
+)
+from forecasting_tools.agents_and_tools.question_generators.question_generator import (
     QuestionGenerator,
+)
+from forecasting_tools.agents_and_tools.question_generators.simple_question import (
     SimpleQuestion,
 )
 from forecasting_tools.ai_models.general_llm import GeneralLlm
